Summary

  Removes the legacy src/core/compute/ engine stack and consolidates distance computation into the scanner under src/core/calc/, split cleanly between Highway (SIMD) and NumKong backends.

  Highlights

  - Removed legacy compute engine (~10k LOC), benchmarks, and tests.
  - Restructured scanner: split ScannerEx into scanner_hw and scanner_nk backends; extracted shared helpers (scanner_dataset_scan, scanner_heap_utils, scanner_query_context, scanner_scan_loops, norm_utils, dist_item).
  - Compile-time compute engine selection replaces the previous runtime stack; legacy config removed.
  - Performance: inlined DataReader::OrderedIterator hot-path accessors, rewrote scan-loop prefetching with a single lookahead advance, made prefetch distance configurable, bitset-filter loop optimization.
  - Docs & tests updated to match; compute_perf tests aligned with new API.

+++ src/core/storage/data_merger.cpp
@@ -1,7 +1,7 @@
 // Implements merge operations for base data files and delta files.
 
 #include "data_merger.h"
-#include "core/compute/compute.h"
+#include "core/calc/norm_utils.h"
 #include "core/storage/data_file_layout.h"
 #include "core/storage/input_reader.h"
 #include "core/storage/compact_ids_ext.h"
@@ -71,9 +71,9 @@ Ret flush_and_close_merge_file(FILE** f, const char* context);
 float compute_stored_norm_for_dist(const uint8_t* data, DataType type, size_t dim, DistFunc dist_func) {
     switch (dist_func) {
         case DistFunc::COS:
-            return compute_cosine_inverse_norm(data, type, dim);
+            return static_cast<float>(compute_cosine_inverse_norm(data, type, dim));
         case DistFunc::L2:
-            return compute_squared_norm(data, type, dim);
+            return static_cast<float>(compute_squared_norm(data, type, dim));
         case DistFunc::DOT:
             throw std::runtime_error("compute_stored_norm_for_dist: DOT does not use stored norms");
         default:
